requirements.txt
streamlit
pandas
scikit-learn
python-Levenshtein
langchain
langchain-core
langchain-groq
groq
python-dotenv